renamed things to make more sense
