--- database_tools/datastores/NumericsRecord.py
@@ -1,8 +1,0 @@
-import wfdb
-from dataclasses import dataclass, InitVar, field
-
-
-@dataclass
-class NumericsRecord:
-    def __post_init__(self):
-        pass
